fix: compute LP trend from match history instead of arbitrary snapshots
The dashboard LP trend was comparing the newest vs oldest of the last 10
rank snapshots, which could all have identical data (snapshots are created
on every sync, even with no new matches), resulting in '+0 LP recently'.

Now computes LP change over the last 10 games by finding the snapshot
closest to the oldest match date and comparing to the latest snapshot.
Also moves the computation server-side and removes the redundant client
IIFE, and drops the 'rank updates tracked' noise text from analytics.
